requirements.txt
flask==3.1.0
authlib==1.3.1
pyjwt==2.8.0
requests==2.31.0
chardet==5.2.0
networkx==3.3